+++ CHANGELOG.md
@@ -21,6 +21,8 @@ When releasing, please add the new-release-boilerplate to docs/pallas/CHANGELOG.
     Please use `jax.lax.pcast(..., to='varying')` as the replacement.
   * `with mesh:` context manager has been deprecated.
     Please use `with jax.set_mesh(mesh):` instead.
+  * Complex arguments passed to {func}`jax.numpy.arange` now result in a
+    deprecation warning, because the output is poorly-defined.
 
 * Changes:
   * jax's `Tracer` no longer inherits from `jax.Array` at runtime. However,